Patch 1 (#1)
* Higher specificity for jsTemplate braces

* Improvements to rest highlighting

* Added higher specificity to the dots in the spread declaration

* Adding jsFuncBlock support

* First pass at handling lamba arguments

* Linking jsArrowFuncArgs to jsFuncArgs

* Allow ... in arrow arguments definition

* Fix Generator definitions

Tacking this on since it relates to ES6 functions.  The generator
astericks were breaking the highlighting of function arguments.  This
should fix generators

* Fix Comment Regions

Our comment detection was a bit buggy because we did not use the extend
keyword.  This meant that if the comment was contained within a region
or a match of sorts, it would get pre-maturely ended by a containing
group.  Using the extend keyword means that it will override any
containing parent matcher and still display a comment, which is how JS
actually interprets the code.

* Improved function declaration destructuring

This should vastly improve how destructuring is handled in function
declarations. It is still far from perfect, and a major problem right
now is how to support multiple lines in lambda functions

* Fix jsBlock vs jsFuncBlock priority

Since we had to remove the contained argument for jsFuncBlock, it caused
false positives for object definitions at the top level of a file.  I've
made jsBlock a higher priorty than jsFuncBlock by placing jsBlock AFTER
jsFuncBlock.

* line following arrow indentation

* Update javascript.vim